feat: :sparkles: editor new validation component
