+++ nuxt-app/components/CodeEditor.vue
@@ -63,7 +63,7 @@ export default {
         tabSize = new Compartment();
 
       let state = EditorState.create({
-        doc: JSON.stringify(self.item.validation, null, 2),
+        doc: JSON.stringify(self.item?.validation, null, 2),
         extensions: [
           basicSetup,
           history(),
