+++ nuxt-app/pages/schema-playground/index.vue
@@ -23,7 +23,7 @@
           />
         </div>
         <div class="col-sm-10 p-2 text-light">
-          <h2 class="font-weight-bold ">
+          <h2 class="font-weight-bold">
             Visualize and Register Your Own Schema
           </h2>
           <p>
@@ -316,7 +316,7 @@ export default {
       number: "",
       loading: false,
       startingPoints: [
-      {
+        {
           namespace: "schema",
           prefix: "schema",
           name: "Dataset",
@@ -329,9 +329,9 @@ export default {
           prefix: "niaid",
           name: "ComputationalTool",
           guide: "/guide/niaid/ComputationalTool",
-          description: "A schema describing a minimal ComputationalTool for the National Institute of Allergy and Infectious Disease (NIAID). A ComputationalTool is a software used for the collection, processing, distribution, analysis, visualization, interpretation, etc. of data. Additional schema.org and/or custom properties could be added.",
+          description:
+            "A schema describing a minimal ComputationalTool for the National Institute of Allergy and Infectious Disease (NIAID). A ComputationalTool is a software used for the collection, processing, distribution, analysis, visualization, interpretation, etc. of data. Additional schema.org and/or custom properties could be added.",
         },
-        
       ],
     };
   },
