+++ nuxt-app/store/modules/editor.js
@@ -1,9 +1,5 @@
 import axios from "axios";
 import Notify from "simple-notify";
-// editor default options
-import { validation_options } from "./editor_options/validation_options";
-import { bioschemas_options } from "./editor_options/bioschemas_options";
-import { definition_options } from "./editor_options/definition_options";
 
 export const editor = {
   state: {
@@ -29,66 +25,13 @@ export const editor = {
     },
     parentDataReceived: false,
     showDescriptions: false,
-    editThis: null,
-    editThisDefinition: null,
     previousPreviewProps: {},
-    validation_options: validation_options,
-    val_test: {
-      name: {},
-      contains_phi: {},
-      description: {},
-      size: {},
-      author: {},
-    },
     top_class_validation: {},
-    definition_options: definition_options,
-    addCardinality: false,
-    toggleValidationOptions: false,
-    bioschemas_options: bioschemas_options
   },
   strict: true,
   mutations: {
     setRemoveValidation(state, payload) {
       state.removeValidation = payload.value;
-    },
-    toggleCardinality(state) {
-      state.addCardinality = !state.addCardinality;
-      //remove all marginality in validation
-      if (!state.addCardinality) {
-        new Notify({
-          status: "warning",
-          title: "Editor",
-          text: "Removing cardinality",
-          effect: "fade",
-          speed: 300,
-          customClass: null,
-          customIcon: null,
-          showIcon: true,
-          showCloseButton: true,
-          autoclose: true,
-          autotimeout: 2000,
-          gap: 20,
-          distance: 20,
-          type: 1,
-          position: "right top",
-        });
-        for (const name in state.finalschema["@graph"][0]["$validation"][
-          "properties"
-        ]) {
-          if (
-            Object.hasOwnProperty.call(
-              state.finalschema["@graph"][0]["$validation"]["properties"][name],
-              "owl:cardinality"
-            )
-          ) {
-            let newV =
-              state.finalschema["@graph"][0]["$validation"]["properties"][name];
-            delete newV["owl:cardinality"];
-            state.finalschema["@graph"][0]["$validation"]["properties"][name] =
-              newV;
-          }
-        }
-      }
     },
     toggleRemoveValidation(state) {
       state.removeValidation = !state.removeValidation;
@@ -137,82 +80,6 @@ export const editor = {
       }
       console.log("✅ Restore Complete!");
     },
-    deleteValidationOption(state, payload) {
-      const msg = [
-        "Adios!",
-        "Bye!",
-        "See Ya!",
-        "Hasta la vista!",
-        "Gone",
-        "Poof!",
-        "So Long!",
-      ];
-      const random = Math.floor(Math.random() * msg.length);
-      for (let i = 0; i < state.validation_options.length; i++) {
-        if (state.validation_options[i]["_id"] == payload.id) {
-          state.validation_options.splice(i, 1);
-          localStorage.setItem(
-            "custom_validation",
-            JSON.stringify(state.validation_options)
-          );
-          new Notify({
-            status: "success",
-            title: "Deleted!",
-            text: msg[random],
-            effect: "fade",
-            speed: 300,
-            customClass: null,
-            customIcon: null,
-            showIcon: true,
-            showCloseButton: true,
-            autoclose: true,
-            autotimeout: 2000,
-            gap: 20,
-            distance: 20,
-            type: 1,
-            position: "right top",
-          });
-        }
-      }
-    },
-    deleteDefinitionOption(state, payload) {
-      const msg = [
-        "Adios!",
-        "Bye!",
-        "See Ya!",
-        "Deleted!",
-        "Gone",
-        "Poof!",
-        "So Long!",
-      ];
-      const random = Math.floor(Math.random() * msg.length);
-      for (let i = 0; i < state.definition_options.length; i++) {
-        if (state.definition_options[i]["_id"] == payload.id) {
-          state.definition_options.splice(i, 1);
-          localStorage.setItem(
-            "custom_definitions",
-            JSON.stringify(state.definition_options)
-          );
-          new Notify({
-            status: "success",
-            title: "Deleted!",
-            text: msg[random],
-            effect: "fade",
-            speed: 300,
-            customClass: null,
-            customIcon: null,
-            showIcon: true,
-            showCloseButton: true,
-            autoclose: true,
-            autotimeout: 2000,
-            gap: 20,
-            distance: 20,
-            type: 1,
-            position: "right top",
-          });
-        }
-      }
-    },
     toggleDesc(state) {
       state.showDescriptions = !state.showDescriptions;
     },
@@ -260,14 +127,6 @@ export const editor = {
       state.finalschema["@graph"][0]["$validation"]["properties"][name] =
         newval;
     },
-    updateValidationOptions(state, payload) {
-      state.validation_options = payload["validation"];
-      // console.log('validation updated from localStorage')
-    },
-    updateDefinitionOptions(state, payload) {
-      state.definition_options = payload["definitions"];
-      // console.log('validation updated from localStorage')
-    },
     resetValidationFor(state, payload) {
       let name = payload["name"];
       let obj = {};
@@ -297,187 +156,6 @@ export const editor = {
         type: 1,
         position: "right top",
       });
-    },
-    addValidationOption(state, payload) {
-      let item = payload["validation"];
-      state.validation_options.push(item);
-      localStorage.setItem(
-        "custom_validation",
-        JSON.stringify(state.validation_options)
-      );
-      new Notify({
-        status: "success",
-        title: "Editor",
-        text: "Option added",
-        effect: "fade",
-        speed: 300,
-        customClass: null,
-        customIcon: null,
-        showIcon: true,
-        showCloseButton: true,
-        autoclose: true,
-        autotimeout: 2000,
-        gap: 20,
-        distance: 20,
-        type: 1,
-        position: "right top",
-      });
-    },
-    addDefinitionOption(state, payload) {
-      function makeID(length) {
-        var result = "";
-        var characters =
-          "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-        var charactersLength = characters.length;
-        for (var i = 0; i < length; i++) {
-          result += characters.charAt(
-            Math.floor(Math.random() * charactersLength)
-          );
-        }
-        return result;
-      }
-
-      function getRandomColor() {
-        var letters = "0123456789ABCDEF";
-        var color = "#";
-        for (var i = 0; i < 6; i++) {
-          color += letters[Math.floor(Math.random() * 16)];
-        }
-        return color;
-      }
-
-      let item = payload["definition"];
-      state.definition_options.push(item);
-      //automatically add validation options with this def
-      state.validation_options.push({
-        _id: makeID(5),
-        title: "(DEF)" + item.title + "(s)",
-        color: getRandomColor(),
-        list: 2,
-        validation: {
-          oneOf: [
-            {
-              $ref: "#/definitions/" + item.title,
-            },
-            {
-              type: "array",
-              items: {
-                $ref: "#/definitions/" + item.title,
-              },
-            },
-          ],
-        },
-        can_delete: true,
-      });
-      state.validation_options.push({
-        _id: makeID(5),
-        title: "(DEF)" + item.title,
-        color: getRandomColor(),
-        list: 2,
-        validation: {
-          $ref: "#/definitions/" + item.title,
-        },
-        can_delete: true,
-      });
-
-      localStorage.setItem(
-        "custom_definitions",
-        JSON.stringify(state.definition_options)
-      );
-      localStorage.setItem(
-        "custom_validation",
-        JSON.stringify(state.validation_options)
-      );
-
-      new Notify({
-        status: "success",
-        title: "Editor",
-        text: item.title + " definition added",
-        effect: "fade",
-        speed: 300,
-        customClass: null,
-        customIcon: null,
-        showIcon: true,
-        showCloseButton: true,
-        autoclose: true,
-        autotimeout: 2000,
-        gap: 20,
-        distance: 20,
-        type: 1,
-        position: "right top",
-      });
-    },
-    editValidationItem(state, payload) {
-      let itemID = payload["item"]["_id"];
-      let newItem = payload["item"];
-      for (let i = 0; i < state.validation_options.length; i++) {
-        let item = state.validation_options[i];
-        if (item["_id"] == itemID) {
-          state.validation_options[i] = newItem;
-          state.editThis = "";
-          new Notify({
-            status: "success",
-            title: "Editor",
-            text: "Edits saved",
-            effect: "fade",
-            speed: 300,
-            customClass: null,
-            customIcon: null,
-            showIcon: true,
-            showCloseButton: true,
-            autoclose: true,
-            autotimeout: 2000,
-            gap: 20,
-            distance: 20,
-            type: 1,
-            position: "right top",
-          });
-        }
-      }
-      localStorage.setItem(
-        "custom_validation",
-        JSON.stringify(state.validation_options)
-      );
-    },
-    editDefinitionItem(state, payload) {
-      let itemID = payload["item"]["_id"];
-      let newItem = payload["item"];
-      for (let i = 0; i < state.definition_options.length; i++) {
-        let item = state.definition_options[i];
-        if (item["_id"] == itemID) {
-          state.definition_options[i] = newItem;
-          state.editThisDefinition = "";
-          new Notify({
-            status: "success",
-            title: "Editor",
-            text: "Edits saved",
-            effect: "fade",
-            speed: 300,
-            customClass: null,
-            customIcon: null,
-            showIcon: true,
-            showCloseButton: true,
-            autoclose: true,
-            autotimeout: 2000,
-            gap: 20,
-            distance: 20,
-            type: 1,
-            position: "right top",
-          });
-        }
-      }
-      localStorage.setItem(
-        "custom_definitions",
-        JSON.stringify(state.definition_options)
-      );
-    },
-    editThis(state, payload) {
-      let item = payload["item"];
-      state.editThis = item;
-    },
-    editThisDefinition(state, payload) {
-      let item = payload["item"];
-      state.editThisDefinition = item;
     },
     savePrefix(state, payload) {
       state.prefix = payload["prefix"];
@@ -976,42 +654,16 @@ export const editor = {
           "https://discovery.biothings.io/view/bioschemas/";
       }
     },
-    toggleValOptions(state) {
-      state.toggleValidationOptions = !state.toggleValidationOptions
-    }
   },
   getters: {
-    addCardinality: (state) => {
-      return state.addCardinality;
-    },
     removeValidation: (state) => {
       return state.removeValidation;
-    },
-    getDefinitionOptions: (state) => {
-      return state.definition_options;
     },
     getSchema: (state) => {
       return state.schema;
     },
-    getEditItem: (state) => {
-      return state.editThis;
-    },
-    getEditDefinitionItem: (state) => {
-      return state.editThisDefinition;
-    },
     getValidationProps: (state) => {
       return state.validation?.["properties"];
-      // return state.val_test
-    },
-    getValidationOptions: (state) => {
-      if (!state.toggleValidationOptions) {
-        return state.validation_options;
-      } else {
-        return state.bioschemas_options;
-      }
-    },
-    getToggleValidationOptions: (state) => {
-      return state.toggleValidationOptions;
     },
     getStartingPoint: (state) => {
       return state.startingPoint;
