+++ nuxt-app/store/modules/editor_options/validation_options.js
@@ -3,12 +3,41 @@
 
 export const validation_options = [
   {
+    _id: "01233bio",
+    title: "text",
+    color: "#097969",
+    validation: {
+      type: "string",
+    },
+    belongs_to: "bioschemas",
+  },
+  {
+    _id: "01432bio",
+    title: "text(s)",
+    color: "#097969",
+    validation: {
+      oneOf: [
+        {
+          type: "string",
+        },
+        {
+          type: "array",
+          items: {
+            type: "string",
+          },
+        },
+      ],
+    },
+    belongs_to: "bioschemas",
+  },
+  {
     _id: "01233",
     title: "string",
     color: "#5D3FD3",
     validation: {
       type: "string",
     },
+    belongs_to: "default",
   },
   {
     _id: "01432",
@@ -27,6 +56,7 @@ export const validation_options = [
         },
       ],
     },
+    belongs_to: "default",
   },
   {
     _id: "14325",
@@ -35,6 +65,7 @@ export const validation_options = [
     validation: {
       type: "boolean",
     },
+    belongs_to: "both",
   },
   {
     _id: "3543",
@@ -43,6 +74,7 @@ export const validation_options = [
     validation: {
       type: "integer",
     },
+    belongs_to: "both",
   },
   {
     _id: "4654",
@@ -52,6 +84,7 @@ export const validation_options = [
       type: "string",
       format: "uri",
     },
+    belongs_to: "both",
   },
   {
     _id: "5867",
@@ -70,6 +103,7 @@ export const validation_options = [
         },
       ],
     },
+    belongs_to: "both",
   },
   {
     _id: "6876",
@@ -79,6 +113,7 @@ export const validation_options = [
       format: "date",
       type: "string",
     },
+    belongs_to: "both",
   },
   {
     _id: "7645",
@@ -99,6 +134,7 @@ export const validation_options = [
         },
       ],
     },
+    belongs_to: "both",
   },
   {
     _id: "8567",
@@ -107,6 +143,7 @@ export const validation_options = [
     validation: {
       const: "Edit const value",
     },
+    belongs_to: "both",
   },
   {
     _id: "9654",
@@ -121,6 +158,7 @@ export const validation_options = [
       },
       required: ["name"],
     },
+    belongs_to: "both",
   },
   {
     _id: "1065",
@@ -138,6 +176,7 @@ export const validation_options = [
       },
       strict: false,
     },
+    belongs_to: "both",
   },
   {
     _id: "1062",
@@ -157,5 +196,6 @@ export const validation_options = [
         },
       ],
     },
+    belongs_to: "both",
   },
 ];
